cambios

+++ src/main/java/com/siapa/dao/DescuentoDao.java
@@ -15,6 +15,6 @@ import org.springframework.stereotype.Repository;
  * @author Joao
  */
 @Repository
-public class DescuentoDao extends GenericDao<Descuento, Long>{
+public class DescuentoDao extends GenericDao<Descuento, Integer>{
     
 }
